refactor(moq-mux): collapse Muxed into Fmp4 and drop convert/
Subscribers no longer have to subscribe the catalog twice or thread a
CatalogConsumer through Muxed. Fmp4::new(BroadcastConsumer) subscribes
the catalog internally, builds the merged init segment from the first
snapshot, and next().await yields the init bytes (once) followed by
moof+mdat fragments in timestamp order across tracks.

Also delete the convert/ module: it had no production callers (only its
own roundtrip test) and shipping 1442 lines of half-baked Legacy↔CMAF
conversion is the kind of dead code we don't want to maintain. Same
behavior is reachable by piping import → export.
