+++ rs/moq-mux/src/lib.rs
@@ -1,7 +1,7 @@
 //! Media muxers and demuxers for MoQ.
 //!
 //! `moq-mux` sits between [`moq_lite`] (the generic pub/sub protocol) and [`hang`]
-//! (the media catalog/container format). It exposes five submodules:
+//! (the media catalog/container format). It exposes four submodules:
 //!
 //! - [`container`]: the wire-level container abstraction and per-track wrappers —
 //!   the [`Container`](container::Container) trait, the [`Hang`](container::Hang) enum
@@ -14,15 +14,12 @@
 //! - [`import`]: pull external media (fMP4, HLS, raw codec bitstreams, …) into a
 //!   moq broadcast — codec demuxers that publish through a
 //!   [`catalog::Producer`].
-//! - [`export`]: subscribe to a moq broadcast and decode media frames —
-//!   [`Muxed`](export::Muxed) merges every track in a broadcast in timestamp order, and
-//!   [`Fmp4`](export::Fmp4) re-encodes decoded frames as ISO-BMFF / CMAF fragments.
-//! - [`convert`]: republish a broadcast in a different container format
-//!   (Legacy ↔ CMAF) without going through an external transcoder.
+//! - [`export`]: subscribe to a moq broadcast and produce media bytes —
+//!   [`Fmp4`](export::Fmp4) yields a single fMP4 / CMAF byte stream (init segment +
+//!   moof+mdat fragments) in timestamp order across tracks.
 
 pub mod catalog;
 pub mod container;
-pub mod convert;
 mod error;
 pub mod export;
 pub mod import;
